Trying to fix build error by relocating locale json files
