+++ packages/design-system/src/components/locale/en.json
