+++ packages/design-system/src/components/locale/es.json
